add test coverage file for custom resource directory
